extract ActiveEffect to its own file
